# receipt-pal-api-doc.yaml
openapi: 3.0.0
paths:
  /receipts:
    post:
      operationId: create-receipt
      summary: ""
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptUrlDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseDto"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ReceiptDTO"
                      error:
                        type: string
                        nullable: true
                        example: null
                        description: An optional error message if something went wrong
    get:
      operationId: get-all-receipts
      summary: ""
      parameters:
        - name: page
          required: true
          in: query
          schema:
            default: 1
            type: integer
        - name: limit
          required: true
          in: query
          schema:
            default: 10
            type: integer
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReceiptDTO"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /receipts/{id}:
    get:
      operationId: get-receipt-by-id
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptDTO"
  /chat/{userId}:
    get:
      operationId: get-chat
      summary: ""
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseDto"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ReceiptDTO"
                      error:
                        type: string
                        nullable: true
                        example: null
                        description: An optional error message if something went wrong
    post:
      operationId: sendMessage
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageDto"
      responses:
        "201":
          description: ""
info:
  title: ReceiptPal
  description: ReceiptPal API description
  version: "1.0"
  contact: {}
tags:
  - name: receipt-pal
    description: ""
servers: []
components:
  schemas:
    ApiResponseDto:
      type: object
      properties:
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - data
        - error
    ItemDTO:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: string
        vatPercentage:
          type: string
        vatAmount:
          type: number
        productType:
          type: string
      required:
        - description
        - quantity
        - unitPrice
        - amount
        - vatPercentage
        - vatAmount
        - productType
    TransactionDetailsDTO:
      type: object
      properties:
        purchasedAt:
          type: string
          format: date-time
        fiscalReceiptNumber:
          type: string
        manufacturingNumber:
          type: string
      required:
        - purchasedAt
        - fiscalReceiptNumber
        - manufacturingNumber
    ReceiptDTO:
      type: object
      properties:
        companyName:
          type: string
        fiscalCode:
          type: string
        address:
          type: string
        registrationNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemDTO"
        totalAmount:
          type: number
        paymentMethod:
          type: string
        transactionDetails:
          $ref: "#/components/schemas/TransactionDetailsDTO"
        receiptId:
          type: string
      required:
        - companyName
        - fiscalCode
        - address
        - registrationNumber
        - items
        - totalAmount
        - paymentMethod
        - transactionDetails
        - receiptId
    ReceiptUrlDTO:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    ChatMessageDto:
      type: object
      properties:
        message:
          type: string
      required:
        - message